Changed GPS parser to the one from AQ -> Now supports nmea, ubx, mtk protocol with automatic baud and protocol selection

+++ FrskyGps/defines.h
@@ -2,11 +2,11 @@
 
 float      heading = 0;
 float      gps_alt = 0;
-float      groundspeed = 0.0f;
+float      groundspeed = 0;
 float      lat = 0;
 float      lon = 0;
 uint8_t    sats = 0;
-uint16_t   fix = 0;
+uint8_t    fix = 0;
 
 byte lat_dir;
 byte lon_dir;
